Moved Configuration model to Tasks
