+++ WebApiProxy.Core/WebApiProxy.Core.csproj
@@ -38,7 +38,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Models\ActionMethodDefinition.cs" />
-    <Compile Include="Models\Configuration.cs" />
     <Compile Include="ConnectionException.cs" />
     <Compile Include="Models\ConstantDefinition.cs" />
     <Compile Include="Models\ControllerDefinition.cs" />
