+++ WebApiProxy.Tasks/Templates/CSharpProxyTemplate.Parameters.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using WebApiProxy.Core.Models;
+using WebApiProxy.Tasks.Models;
 
 
 namespace WebApiProxy.Tasks.Templates
