+++ WebApiProxy.Tasks/WebApiProxy.Tasks.csproj
@@ -55,6 +55,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Models\Configuration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ProxyGenerationTask.cs" />
     <Compile Include="Templates\CSharpProxyTemplate.cs">
